+++ src/docs/devguide/hugo/site/content/release-notes.md
@@ -1,12 +1,28 @@
-## v1.1.2 
+## v2.0.0
+Released 25 Oct 2016
+
+### Breaking changes
+* Introduced `PurchaseRequestRetry` model to be used in place of `PurchaseRequest` for request retries
+* Introduced `SlipData` model to be used in place of `PrintableSlip`
+* Removed `slipWdth` from requests as this is now contained in `SlipData`
+* Introduced `ConfirmationAdvice` and `ReversalAdvice` models as wrappers for `TenderAdvice` and `BasicReversal`, respectively
+* Refactored `ErrorDetail` model
+
+### Non-breaking changes
+* Added `msisdn` and `tenders` fields to `PurchaseRequest` model
+* Added `BasicAdviceResponse` body to `reverseTokenPurchase` and `confirmTokenPurchase` operations
+* Added `LOYALTY_CARD` and `GIFT_CARD` tender types
+
+
+## v1.1.2
 Released 19 Sep 2016
 
 * Bug fix to JSON validator
 
-## v1.1.1 
+## v1.1.1
 Released 15 Sep 2016
 
-* Bug fix to TaxableAmount getter methods
+* Bug fix to `TaxableAmount` getter methods
 
 ## v1.1.0
 Released 12 Sep 2016